refactor: 멤버 도메인 import 변경 #131
