+++ src/main/java/com/bookripple/api/domain/member/exception/DuplicateEmailException.java
@@ -1,12 +1,10 @@
 package com.bookripple.api.domain.member.exception;
 
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.error.ApiException;
+import com.bookripple.api.global.code.CommonErrorCode;
+import com.bookripple.api.global.error.ApiException;
+
 /**
- * 이메일 중복 시 발생
- * - 회원가입
- * - 이메일 변경
- * - 소셜 로그인 연동 시
+ * 이메일 중복 시 발생 - 회원가입 - 이메일 변경 - 소셜 로그인 연동 시
  */
 public class DuplicateEmailException extends ApiException {
