+++ src/main/java/com/bookripple/api/domain/member/exception/DuplicateLoginIdException.java
@@ -1,11 +1,10 @@
 package com.bookripple.api.domain.member.exception;
 
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.error.ApiException;
+import com.bookripple.api.global.code.CommonErrorCode;
+import com.bookripple.api.global.error.ApiException;
+
 /**
- * 로그인 아이디 중복 시 발생
- * - 회원가입
- * - 아이디 변경
+ * 로그인 아이디 중복 시 발생 - 회원가입 - 아이디 변경
  */
 public class DuplicateLoginIdException extends ApiException {
